+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.0.2
 - Added quick-links so users can visit a site easily
+- Updated generic-components
 
 ## 1.0.1
 - Fixed bug in keyboard control for dark mode toggle